requirements.txt
anthropic>=0.18.0
openai>=1.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.2
PyPDF2>=3.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
surya-ocr>=0.14.6
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.35.0
